// libs/features/collection/billing/src/lib/shared/enums/validator-roles.ts
export enum ValidatorRoles {
  VALIDATOR = 'Validator',
  VALIDATOR_ONE = 'Validator1',
  VALIDATOR_TWO = 'Validator2',
  FC_VALIDATOR = 'FCApprover',
  GDIC = 'GDIC',
  GDISO = 'GDISO',
  GDES = 'GeneralDirectorforEstablishmentServices',
  GDISOFULLNAME = 'GeneralDirectorofInsuranceServicesandOperations',
  DOCTOR = 'Doctor',
  CDM = 'Collection Department Manager',
  COLLECTION_MANAGER = 'colldepmgr',
  AG = 'AG',
  GOVERNOR = 'Governor'
}